change names components
