+++ src/components/ChatItem.jsx
@@ -1,44 +1,17 @@
-import React, { Fragment } from 'react';
-import classNames from 'classnames';
-import { withStyles } from 'material-ui/styles';
-import Typography from 'material-ui/Typography';
-import Paper from '@material-ui/core/Paper';
+import React, { Component } from 'react';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemText from '@material-ui/core/ListItemText'
+import Avatar from 'material-ui/Avatar';
 
-const styles = theme => ({
-  message: {
-    maxWidth: '70%',
-    minWidth: '10%',
-    padding: theme.spacing.unit,
-    marginLeft: theme.spacing.unit * 2,
-  },
-  messageFromMe: {
-    marginRight: theme.spacing.unit * 2,
-    backgroundColor: '#e6dcff'
-  },
-});
+import titleInitials from '../utils/title-initials'
 
-const ChatItem = props => {
-  const {
-    classes,
-    message,
-    isMessageFromMe,
-    userAvatar
-  } = props;
-
+const ChatItem = ({ chat }) => {
   return (
-    <Fragment>
-      {!isMessageFromMe && userAvatar}
-      <Paper className={classNames(classes.message, isMessageFromMe && classes.messageFromMe)}>
-        <Typography variant="caption">
-          {message.sender}
-        </Typography>
-        <Typography variant="body1">
-          {message.content}
-        </Typography>
-      </Paper>
-      {isMessageFromMe && userAvatar}
-    </Fragment>
-  )
+    <ListItem button>
+      <Avatar>{titleInitials(chat.title)}</Avatar>
+      <ListItemText primary={chat.title}/>
+    </ListItem>
+  );
 }
 
-export default withStyles(styles)(ChatItem);
+export default ChatItem;
