+++ src/index.js
@@ -2,6 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'typeface-roboto';
 import './index.css';
-import App from './App';
+import App from './components/App.jsx';
 
 ReactDOM.render(<App />, document.getElementById('root'));
